UPSTREAM: drm/i915/ehl: Ensure that the DDI selection MUX is programmed correctly
Perhaps in some cases the BIOS/GOP or other firmware may turn on
PHY A but may not program the MUX correctly. Therefore, re-program
PHY A if it is determined after reading the VBT that the value
programmed for the MUX bit does not match the expected value.

Cc: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200121235848.8457-1-vivek.kasireddy@intel.com
(cherry picked from commit 646603d9aba7c360ff9e5201bec479219c0a0a7e)



BUG=b:168131866
TEST=Compile tested/Backported fixes identified with r-o-m/On-device testing with eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I84fb79992d99f96e2fb4d63323525911078c9b98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2405137
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>