ANDROID: selinux: modify RTM_GETLINK permission
This is a cherrypick of http://r.android.com/1215021

Map the permission gating RTM_GETLINK messages to a new permission so
that it can be distinguished from the other netlink route permissions
in selinux policy.

This is a temporary Android-only patch that will be deprecated in
newer kernels once the long-term solution lands as discusssed on the
mailing list [1]. The maintainer's recommended solution is more
general, much more complex, and likely not suitable for backporting.
This patch provides the minimal change needed for Android including
the userspace settable trigger which ensures that the permission
change is only applied to the newest version of Android which
contains the changes needed for userpace compatibility.

[1]:
https://lore.kernel.org/selinux/20200116142653.61738-1-jeffv@google.com/

BUG=b:158561720
TEST=atest android.security.SELinuxTargetSdkTest#testNoNetlinkRouteGetlink

Change-Id: I7b44ce60ad98f858c412722d41b9842f8577151f
Signed-off-by: Jeff Vander Stoep <jeffv@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2241414
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Satoshi Niwa <niwa@chromium.org>
Tested-by: Satoshi Niwa <niwa@chromium.org>