UPSTREAM: dt-bindings: fix warnings in validation of qcom,gcc.yaml
The last example in qcom,gcc.yaml set 'sleep' as the second
value of 'clock-names'. According to the schema is should
be 'sleep_clk'. Fix the example to conform the schema.
This fixes a warning when validating the schema:
"clock-names:  ... is not valid under any of the given schemas"

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Link: https://lkml.kernel.org/r/20200122135741.12123-1-dafna.hirschfeld@collabora.com
Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit 7914d5c6266abff4918f0fd91ee53e8b537d65a1)

BUG=b:143638832
TEST=None

Change-Id: Ifef4f43eb0e625f5b196b4b1604d55ad750674ed
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2037460
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>