CHROMIUM: iwl7000: mac80211: consider more elements in parsing CRC
We only use the parsing CRC for checking if a beacon changed,
and elements with an ID > 63 cannot be represented in the
filter. Thus, like we did before with WMM and Cisco vendor
elements, just statically add these forgotten items to the
CRC:
 - WLAN_EID_VHT_OPERATION
 - WLAN_EID_OPMODE_NOTIF
 - WLAN_EID_EXT_HE_OPERATION.

I guess that in most cases when VHT/HE operation change, the HT
operation also changed, and so the change was picked up, but we
did notice that pure operating mode notification changes were
ignored.

Change-Id: Ibc63a8ca05014793fb6bf567da2a3746472514b8
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4da69f421b332f811e81103dcad31dcbcb4e4578