FROMLIST: drm/amd/display: Only enable cursor on pipes that need it
[Why]
In current code we're essentially drawing the cursor on every pipe
that contains it. This only works when the planes have the same
scaling for src to dest rect, otherwise we'll get "double cursor" where
one cursor is incorrectly filtered and offset from the real position.

[How]
Without dedicated cursor planes on DCN we require at least one pipe
that matches the scaling of the current timing.

This is an optimization and workaround for the most common case where
the top-most plane is not scaled but the bottom-most plane is scaled.

Whenever a pipe has a parent pipe in the blending tree whose recout
fully contains the current pipe we can disable the pipe.

This only applies when the pipe is actually visible of course.

Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
(am from https://patchwork.freedesktop.org/series/72789/)

 Conflicts:
	drivers/gpu/drm/amd/display/dc/dcn10/dcn10_hw_sequencer.c
 [shirish.s: fixed trivial rebase conflicts that occurred due to
 skipping intermediate non-important patches.]

BUG=b:145260834
TEST=No 2 cursors are visible with overlay enabled,
     while running https://webglsamples.org/aquarium/aquarium.html demo

Change-Id: I9205ce34d81294987d808e4863a9e41bc8445b16
Signed-off-by: Shirish S <shirish.s@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2032542
Reviewed-by: Raul E Rangel <rrangel@chromium.org>