UPSTREAM: Documentation: admin-guide: PM: Update sleep states documentation
There is some information in Documentation/power/interface.rst that
is still missing from Documentation/admin-guide/pm/sleep-states.rst
and really should be present in there, so update the latter by
adding that information to it and delete the former (as it becomes
redundant after that and it is somewhat outdated).

While at it, clean up some assorted pieces of sleep-states.rst a bit.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit c21502efdaedfdf9fc71334883a164341881bc22)

BUG=b:146144745
TEST=Kernel boots on volteer

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I3f752b68194206fee485834160620d97d96e7c89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2092349
Reviewed-by: Alex Levin <levinale@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>