UPSTREAM: intel_idle: Introduce 'use_acpi' module parameter
For diagnostics, it is generally useful to be able to make intel_idle
take the system's ACPI tables into consideration even if that is not
required for the processor model in there, so introduce a new module
parameter, 'use_acpi', to make that happen and update the documentation
to cover it.

While at it, fix the 'no_acpi' module parameter name in the
documentation.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 3a5be9b8f43346a24f31c0017cb2566a6b2c72c5)

BUG=b:146144745
TEST=Kernel boots on volteer

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: Id825c3d2d971abb950bb5d541a3ff2f4d43416a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2092347
Reviewed-by: Alex Levin <levinale@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>