CHROMIUM: iwl7000: Revert "chromeOS: update ifdef for cfg80211_send_layer2_update() definition"
This reverts commit 055b72e161492da1dfade98843577773e3b3b06c.

In ChromeOS, the original patch in cfg80211 that was implementing this
was reverted, so we need to revert the change here as well.

I have sent a patch to ChromeOS to add the cfg80211 implementation
back and our ifdef change at the same time, to avoid build breakages.
When that patch gets applied, we need to revert this revert in our
tree too.

Change-Id: Iacacf03838c0e38353cc04571a97305bbd877348
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: cd3ee683e1e3e5560d7e5f8a5b6fe3a70392ea72