UPSTREAM: of/address: Support multiple 'dma-ranges' entries
Currently, the DMA offset and mask for a device are set based only on the
first 'dma-ranges' entry. We should really be using all the entries. The
kernel doesn't yet support multiple offsets and sizes, so the best we can
do is to find the biggest size for a single offset. The algorithm is
copied from acpi_dma_get_range().

If there's different offsets from the first entry, then we warn and
continue. It really should be an error, but this will likely break
existing DTs.

Signed-off-by: Rob Herring <robh@kernel.org>
(cherry picked from commit 9d55bebd9816903b821a403a69a94190442ac043)

BUG=b:187155802
TEST=Boot OK

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: Ibd0d25f4c3dd1989b5252172fc68237d5e26e4d3
Signed-off-by: Flora Fu <flora.fu@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2738869
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>