BACKPORT: x86/vdso: Move VDSO clocksource state tracking to callback
All architectures which use the generic VDSO code have their own storage
for the VDSO clock mode. That's pointless and just requires duplicate code.

X86 abuses the function which retrieves the architecture specific clock
mode storage to mark the clocksource as used in the VDSO. That's silly
because this is invoked on every tick when the VDSO data is updated.

Move this functionality to the clocksource::enable() callback so it gets
invoked once when the clocksource is installed. This allows to make the
clock mode storage generic.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Michael Kelley <mikelley@microsoft.com>  (Hyper-V parts)
Reviewed-by: Vincenzo Frascino <vincenzo.frascino@arm.com> (VDSO parts)
Acked-by: Juergen Gross <jgross@suse.com> (Xen parts)
Link: https://lkml.kernel.org/r/20200207124402.934519777@linutronix.de

(cherry picked from commit eec399dd862762b9594df3659f15839a4e12f17a)

Conflicts:
   arch/x86/entry/vdso/vma.c
   drivers/clocksource/hyperv_timer.c

BUG=b:151181951
TEST=vdso works on arm32 userspace

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I80163c8a0a037984aed34f3f0588658a4c8cbf75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2154133
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>