CHROMIUM: iwl7000: mac80211: Fail association when AP has no legacy rates
The MLME logic had a workaround that allowed to continue an
association with an AP even if the AP did not provide any basic
rates in its supported rates in the association response, assuming
that the first (non basic) legacy rate could be used as a basic rate.
However, this did not consider the case where the AP (which is
obviously buggy) did not provide any legacy rate.

Fix this by failing the association, as this can result in
an unexpected failure in the low level driver and FW, e.g., in
rate scale logic etc.

Change-Id: I6e6ce5efda351a8544c0e7bfeee260fe3360d401
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 13125d79156c72b1e7d71d4aaf25845730d614f9