UPSTREAM: PM: QoS: Clean up pm_qos_read_value() and pm_qos_get/set_value()
Move the definition of pm_qos_read_value() before the one of
pm_qos_get_value() and add a kerneldoc comment to it (as it is
not static).

Also replace the BUG() in pm_qos_get_value() with WARN() (to
prevent the kernel from crashing if an unknown PM QoS type is
used by mistake) and drop the comment next to it that is not
necessary any more.

Additionally, drop the unnecessary inline modifier from the header
of pm_qos_set_value().

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>
Tested-by: Amit Kucheria <amit.kucheria@linaro.org>
(cherry picked from commit dcd70ca1a3bf33aa5cf53aa6f72b8e51afb1ac48)

BUG=b:168669388
TEST=Kernel compiles and boots.
     emerge-volteer chromeos-kernel-5_4
     ~/trunk/src/scripts/update_kernel.sh --remote ${DUT_IP}
TEST=Tast spot checks pass.
     tast run ${DUT_IP} arc.Boot graphics.GLBench.hasty \
                        ui.ChromeLogin ui.TabSwitchCUJ \
                        video.Play.vp9 video.Seek.stress_h264 \
                        webrtc.RTCPeerConnection.vp8

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I6e86ba6a7c69625116bc8488544a91a7ff0f003d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2462442
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>