FROMGIT: thunderbolt: Disable lane 1 for XDomain connection
USB4 spec mandates that the lane 1 should be disabled if lanes are not
bonded. For host-to-host connections (XDomain) we don't support lane
bonding so in order to be compatible with the spec, disable lane 1 when
another host is connected.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 341d45188a7800ae3bc18558d62020787b78397e
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:162700590
TEST=No regressions for the following use cases:
- S0iX NDA/DA
- TBT NVMe hotplug
- TBT dock hotplug
- TBT dock w/ USB pen drive
- TBT dock w/ display

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I80e0c59e58adfc4f98a2f91c19529325c27da4b4