UPSTREAM: iio: core: drop devm_iio_device_free() API call
It's unused so far, so it can be removed. Also makes sense to remove it
to discourage weird uses of this call during review.

This is the last user of 'devm_iio_device_match()', so it can be removed as
well in this patch.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
(cherry picked from commit 66be392a48f9d5256b4ba582ff06303fe97b5ca2)
Signed-off-by: Daniel Campello <campello@chromium.org>

BUG=b:157206958,b:111806212,b:162279829
TEST=emerge-volteer chromeos-kernel-5_4

Change-Id: I3a1404d90e4864d5562f75a260df94151718e223
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2351181
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>