UPSTREAM: drm/i915/gt: Expose preempt reset timeout via sysfs
After initialising a preemption request, we give the current resident a
small amount of time to vacate the GPU. The preemption request is for a
higher priority context and should be immediate to maintain high
quality of service (and avoid priority inversion). However, the
preemption granularity of the GPU can be quite coarse and so we need a
compromise.

The preempt timeout can be adjusted per-engine using,

	/sys/class/drm/card?/engine/*/preempt_timeout_ms

and can be disabled by setting it to 0.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Steve Carbonari <steven.carbonari@intel.com>
Tested-by: Steve Carbonari <steven.carbonari@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200228131716.3243616-6-chris@chris-wilson.co.uk
(cherry picked from commit db3d8338bacbfc734d246eaccfbe5b19ad81d7d4)
Signed-off-by: Patrick Waterlander <patrick.waterlander@intel.com>

BUG=b:162528275
TEST=glbench test passes on Volteer with no GPU hangs

Change-Id: I3d788ddcab16a76f9a0b0c34f8b41d3ddd2b1695
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2360443
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Tested-by: Patrick Waterlander <patrick.waterlander@intel.com>
Commit-Queue: Patrick Waterlander <patrick.waterlander@intel.com>